Move ThrowableProxyComponent logic into Activator Class and remove the ThrowableProxyComponent class
