+++ dev/com.ibm.ws.logging.osgi/bnd.bnd
@@ -37,8 +37,8 @@ Service-Component= \
   com.ibm.ws.logging.internal.ffdcjanitor; \
     scheduler=java.util.concurrent.ScheduledExecutorService; \
     implementation:=com.ibm.ws.logging.internal.osgi.FFDCJanitor, \
-  com.ibm.ws.logging.internal.osgi.stackjoiner.ThrowableProxyComponent; \
-    implementation:="com.ibm.ws.logging.internal.osgi.stackjoiner.ThrowableProxyComponent"; \
+  com.ibm.ws.logging.internal.osgi.Activator; \
+    implementation:="com.ibm.ws.logging.internal.osgi.Activator"; \
     instrumentation=java.lang.instrument.Instrumentation; \
     properties:="service.vendor=IBM"; \
     version:=1.1 \
