+++ dev/com.ibm.ws.logging.osgi/src/com/ibm/ws/logging/internal/osgi/Activator.java
@@ -17,11 +17,13 @@ import org.eclipse.equinox.log.ExtendedLogReaderService;
 import org.osgi.framework.BundleActivator;
 import org.osgi.framework.BundleContext;
 import org.osgi.framework.ServiceReference;
+import org.osgi.service.component.ComponentContext;
 import org.osgi.service.log.LogReaderService;
 
 import com.ibm.websphere.ras.TrConfigurator;
 import com.ibm.websphere.ras.TraceComponent;
 import com.ibm.websphere.ras.TraceComponentChangeListener;
+import com.ibm.ws.logging.internal.osgi.stackjoiner.ThrowableProxyActivator;
 import com.ibm.ws.ras.instrument.internal.main.LibertyJava8WorkaroundRuntimeTransformer;
 import com.ibm.ws.ras.instrument.internal.main.LibertyRuntimeTransformer;
 
@@ -40,7 +42,42 @@ public class Activator implements BundleActivator {
     private RuntimeTransformerComponentListener runtimeTransformer = null;
 
     private static TraceComponentChangeListenerTracker listenerTracker = null;
+    private ComponentContext componentContext;
+	private ThrowableProxyActivator proxyActivator;
+	
+	/**
+     * Activation callback from the Declarative Services runtime where the
+     * component is ready for activation.
+     *
+     * @param bundleContext the bundleContext
+     */
+    synchronized void activate(ComponentContext componentContext) throws Exception {
+    	this.componentContext = componentContext;
+        this.proxyActivator = new ThrowableProxyActivator(this.inst, componentContext.getBundleContext());
+        this.proxyActivator.activate();
+    }
+    
+    /**
+     * Deactivation callback from the Declarative Services runtime where the
+     * component is deactivated.
+     *
+     * @param bundleContext the bundleContext
+     */
+    synchronized void deactivate() throws Exception {
+        this.proxyActivator.deactivate();
+    }
 
+
+    /**
+     * 
+     * Inject reference to the {@link java.lang.instrument.Instrumentation} implementation.
+     *
+     * @param instrumentationAgent the JVM's {@code Instrumentation) reference
+     */
+    protected void setInstrumentation(Instrumentation instrumentation) {
+        this.inst = instrumentation;
+    }
+    
     /**
      * @see org.osgi.framework.BundleActivator#start(org.osgi.framework.BundleContext)
      * @see com.ibm.liberty.kernel.internal.Activator#start(org.osgi.framework.BundleContext)
